Repair QEMU test fully.
The test showed that the git-hook-impls symlink may not in fact work with older
hooks as intended.  This is because the old call to git-archive only copied over
the symlink itself and not the target.  Therefore the symlink was not sufficient
to preserve complete backward compatibility, though it is still slightly useful.